Added scalar projection methods to CatmullRomCurve
Recursive methods are used to determine which vertices along the curve should have lines drawn between them, balancing cost and appearance. A binary search method is used to determine the t value of the closest point along the curve to an arbitrary point, and a method that calculates the curve's length is used to determine the binary search's depth.