Added algorithm for finding closest point on complex path
It works amazinglygit add .git add . It can work for any shape of path, and looks especially cool on curvy paths. It uses a divide and conquer approach where it checks which individual curve it is closest to, where that point is on the individual curve, and what its t value would be along the entire path. Next step is to add vehicles.